api/requirements.txt
Django==2.1
django-rest-framework==0.1.0
djangorestframework==3.8.2